Refactor unroll command and enhance error handling
- Updated `unroll` command in `main.py` to accept optional parameters for `skip_files` and `overwrite`.
- Introduced a new method in `QasmModule` to determine if files should be skipped based on specific tags.
- Enhanced error reporting in the `unroll_qasm` function to include raw logger output for better debugging.
- Added comprehensive tests for the `unroll` command, covering various scenarios including file skipping, output handling, and error cases.
